+++ src/interactive/repl.ts
@@ -766,6 +766,9 @@ async function executeCell(shouldMove: boolean = false) {
     if (ed === undefined) {
         return
     }
+    if (vscode.workspace.getConfiguration('julia').get<boolean>('execution.saveOnEval') === true) {
+        await ed.document.save()
+    }
 
     const doc = ed.document
     const selection = ed.selection
@@ -796,7 +799,9 @@ async function evaluateBlockOrSelection(shouldMove: boolean = false) {
     if (editor === undefined) {
         return
     }
-
+    if (vscode.workspace.getConfiguration('julia').get<boolean>('execution.saveOnEval') === true) {
+        await editor.document.save()
+    }
     const selections = editor.selections.slice()
 
     await startREPL(true, false)